+++ src/components/index.js
@@ -3,4 +3,6 @@ export Button from './Button';
 export ColorSelector from './ColorSelector';
 export Dropdown from './Dropdown';
 export OverflowMenu from './OverflowMenu';
+export Icon from './Icon';
+export Table from './Table';
 export Searchbar from './Searchbar';